+++ Tests/MongoSwiftTests/AuthTests.swift
@@ -2,6 +2,7 @@ import Foundation
 import mongoc
 @testable import MongoSwift
 import Nimble
+import TestsCommon
 import XCTest
 
 /// An extension adding accessors for a number of options that may be set on a `ConnectionString`.
