+++ Tests/MongoSwiftTests/ClientSessionTests.swift
@@ -1,6 +1,7 @@
 import Foundation
 @testable import MongoSwift
 import Nimble
+import TestsCommon
 import XCTest
 
 final class ClientSessionTests: MongoSwiftTestCase {
