+++ Tests/MongoSwiftTests/CommandMonitoringTests.swift
@@ -1,5 +1,6 @@
 @testable import MongoSwift
 import Nimble
+import TestsCommon
 import XCTest
 
 let center = NotificationCenter.default
