+++ Tests/MongoSwiftTests/MongoClientTests.swift
@@ -2,6 +2,7 @@ import Foundation
 import mongoc
 @testable import MongoSwift
 import Nimble
+import TestsCommon
 import XCTest
 
 final class MongoClientTests: MongoSwiftTestCase {
