+++ Tests/MongoSwiftTests/MongoCollection+BulkWriteTests.swift
@@ -1,5 +1,6 @@
 @testable import MongoSwift
 import Nimble
+import TestsCommon
 import XCTest
 
 final class MongoCollection_BulkWriteTests: MongoSwiftTestCase {
