+++ Tests/MongoSwiftTests/MongoCursorTests.swift
@@ -1,6 +1,7 @@
 import Foundation
 @testable import MongoSwift
 import Nimble
+import TestsCommon
 import XCTest
 
 private let doc1: Document = ["_id": 1, "x": 1]
