+++ Tests/MongoSwiftTests/ReadWriteConcernTests.swift
@@ -1,6 +1,7 @@
 import mongoc
 @testable import MongoSwift
 import Nimble
+import TestsCommon
 import XCTest
 
 extension WriteConcern {
